Added CLI
Added CLI to the deploy script. Might need a better name than deploy but
it's fine for now. Just has capability to deploy and delete right now.
Might want to add subcommand for suspending the instance without
deleting everything.

Also added some helpful print commands to the bootstrap script to see
what stage the bootstrap is in on stdout