Extended CLI
Added stop and start subcommands, to start or stop instances within a
stack. Useful for saving money without completely terminating an
instance and thus deleting all the data within it and forcing a complete
redeploy/bootstrap, which can take awhile.

TODO:

It might be nice to include the status of the instance within a stack,
when listing all running stacks. THis way users won't have to remember
which stacks have a running instance and which stacks don't. Not totally
necessary, but would be a nice feature.